TW-37477 don't publish status for queued build
A queued build can be optimized away if there is a started build on the
same revision or there is a build with more recent changes. In first
case we can reuse the status of the replacement build, in the latter
case we cannot, since replacement build contains more changes.

Since Stash doesn't allow to remove a build status it is better to
publish status only for started builds as we do in other publishers.